Refactored logic to break a poem down to stanzas & lines to helpers.py
